refactor: 이벤트 바인딩과 핸들러 구조 정리
