+++ src/view/movieDetail.ts
@@ -1,8 +1,9 @@
+import { RATING_SCORES, RATING_TEXTS } from "../constants/rating";
 import { MovieDetail } from "../services/dto";
 
 export const renderMovieDetail = (movieDetail: MovieDetail) => {
   const movieModal = document.querySelector<HTMLElement>(".modal");
-  if (!movieModal) return null;
+  if (!movieModal) return;
 
   movieModal.dataset.movieId = String(movieDetail.id);
 
@@ -13,12 +14,11 @@ export const renderMovieDetail = (movieDetail: MovieDetail) => {
   const rate = movieModal.querySelector(".rate-value");
   const detail = movieModal.querySelector(".detail");
 
-  if (modalImage)
+  if (modalImage) {
     modalImage.src =
       `https://media.themoviedb.org/t/p/w300_and_h450_face` +
       movieDetail.poster_path;
-
-  if (title) title.textContent = movieDetail.title;
+  }
 
   if (category) {
     const releaseYear = new Date(movieDetail.release_date).getFullYear();
@@ -26,15 +26,16 @@ export const renderMovieDetail = (movieDetail: MovieDetail) => {
     category.textContent = `${releaseYear} · ${genres}`;
   }
 
+  if (title) title.textContent = movieDetail.title;
   if (rate) rate.textContent = movieDetail.vote_average.toString();
-
   if (detail) detail.textContent = movieDetail.overview;
 
-  const key = window.localStorage.getItem(String(movieDetail.id));
-  if (!key) return;
+  const ratingScore = window.localStorage.getItem(String(movieDetail.id));
+  if (!ratingScore) return;
 
-  const stars = movieModal.querySelectorAll<HTMLImageElement>(".stars img");
-  stars[Number(key) / 2 - 1].click();
+  const index = RATING_SCORES.indexOf(ratingScore);
+  fillStars(index);
+  updateRatingResult(index);
 };
 
 export const clearMovieDetail = () => {
@@ -49,9 +50,28 @@ export const clearMovieDetail = () => {
   const stars = movieModal.querySelectorAll<HTMLImageElement>(".stars img");
   stars.forEach((star) => (star.src = "./images/star_empty.png"));
 
-  const ratingText = document.querySelector(".rating-text");
+  const ratingText = movieModal.querySelector(".rating-text");
   if (ratingText) ratingText.textContent = "평가해주세요";
 
+  const ratingValue = movieModal.querySelector("#rating-value");
+  if (ratingValue) ratingValue.textContent = "0";
+};
+
+export const fillStars = (index: number) => {
+  const stars = document.querySelectorAll<HTMLImageElement>(".stars img");
+
+  stars.forEach((currentStar, currentIndex) => {
+    currentStar.src =
+      currentIndex <= index
+        ? "./images/star_filled.png"
+        : "./images/star_empty.png";
+  });
+};
+
+export const updateRatingResult = (index: number) => {
+  const ratingText = document.querySelector(".rating-text");
+  if (ratingText) ratingText.textContent = RATING_TEXTS[index];
+
   const ratingValue = document.querySelector("#rating-value");
-  if (ratingValue) ratingValue.textContent = (0).toString();
+  if (ratingValue) ratingValue.textContent = RATING_SCORES[index];
 };
